+++ src/test/resources/ru/vyarus/gradle/plugin/animalsniffer/java/skip/Sample.java
@@ -1,0 +1,13 @@
+package invalid;
+
+public class Sample {
+
+    public static void main(String[] args) {
+        // method added in 1.7
+        Boolean.compare(true, true);
+    }
+
+    public void someth() {
+
+    }
+}