+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Add optional incremental mode (#4)
+
 ### 1.3.0 (2017-03-20)
 * Prevent other tasks output interception (#3)
 * Change console reporting format (to match gradle-quality-plugin format):